Updated Navigation Bar
Changed the navigation bar to be lower on the screen. Included icons for each of the navigation bar pages which is found in a new folder titled "images" (MuHASS-App\src\images). Included a feature to highlight the current tab the user is on. Currently there isn't an active home tab. This will be replaced by the connectivity page which will show all available devices able to connect to the app (bluetooth page).